requirements.txt
Django==4.2.3
python-dotenv==1.0.0
djangorestframework==3.14.0
Pillow==10.0.0
django-phonenumber-field==7.1.0
phonenumbers==8.13.17
drf-yasg2==1.19.4